feat: Add Member Center view and navigation
Introduces a new `MEMBER_CENTER` view state to the application. This change also updates the Navbar to conditionally navigate to the `MEMBER_CENTER` view when the user icon is clicked, instead of defaulting to the home page if the user is logged in. The associated UI elements and translations for the member login and related features are also prepared.